fix: controls
